GRUPO COLON - MASTER PROYECTOS I.T

+++ ejercicio13_ejemplo_tablas/ejercicio13_entrega/ejercicio13_entrega.py
@@ -59,7 +59,7 @@ for i in range(len(datos)):
         suma += edad
         cantidad += 1
 
-    media = suma / cantidad
+    media = int(suma / cantidad)
 
 print ("La edad media de los alumnos es: ", media)
